+++ tool/frame.rkt
@@ -7,15 +7,19 @@
          racket/gui/base
 
          sauron/project/manager
-         sauron/project/current-project
-         sauron/project/panel)
+         sauron/project/panel
+         sauron/log)
 
 (define tool@
   (unit
     (import drracket:tool^)
     (export drracket:tool-exports^)
 
-    (define (phase1) (void))
+    (define (phase1)
+      (preferences:set-default 'current-project #f (λ (_) #t))
+      (preferences:add-callback 'current-project
+                                (λ (_ new-dir)
+                                  (log:info "current project is ~a" new-dir))))
     (define (phase2) (void))
 
     (define drracket-frame-mixin
@@ -52,10 +56,10 @@
                          [label "select a project"]
                          [on-select
                           (λ (path)
-                            (send current-project set path)
+                            (preferences:set 'current-project path)
                             (show-real-area)
                             (send c set-label "Hide the Project Viewer"))]))
-                  (if (send current-project get)
+                  (if (preferences:get 'current-project)
                       (if project-files-show?
                           (let ()
                             (close-real-area)